Split components of SelectedScheduleCard
